refactored lyft code
